+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.0.3
+- Music now loops
+- Added visualization of block holding
+- Small changes to blockdata structure
+- Added ingame commands (t)
+- Started working on inventory system
+
 # 0.0.2
 - Added ingame music
 - Removed/Moved some world generator code (Now doing less math per chunk!)